Move `ImageCommentsUIIntegrationTests` helper methods to separate files
