+++ EssentialApp/EssentialAppTests/ImageComments Helpers/ImageCommentsUIIntegrationTests+Localization.swift
@@ -1,0 +1,23 @@
+//
+//  ImageCommentsUIIntegrationTests+Localization.swift
+//  EssentialAppTests
+//
+//  Created by Adrian Szymanowski on 16/03/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+import XCTest
+import EssentialFeed
+
+extension ImageCommentsUIIntegrationTests {
+	func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
+		let table = "ImageComments"
+		let bundle = Bundle(for: ImageCommentsPresenter.self)
+		let value = bundle.localizedString(forKey: key, value: nil, table: table)
+		if value == key {
+			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
+		}
+		return value
+	}
+}
